Refactored stan sampling procedure into sampling/ dir from solvers/
